Update main.py
I tried to see if it would go into the subdirectory folder to find the file If I added it to the path name. 
It seems like the initial "this_file_ path" is reading where this file is currently, then this is telling the other line to find the files based on where this file is.